Split SQS receive request instrumentation so we can apply structural constraint to the QueueBufferConfig part
