+++ dd-java-agent/instrumentation/aws-java-sqs-1.0/src/main/java/datadog/trace/instrumentation/aws/v1/sqs/SqsReceiveRequestInstrumentation.java
@@ -1,10 +1,7 @@
 package datadog.trace.instrumentation.aws.v1.sqs;
 
-import static datadog.trace.agent.tooling.bytebuddy.matcher.NameMatchers.named;
 import static datadog.trace.agent.tooling.bytebuddy.matcher.NameMatchers.namedOneOf;
-import static java.util.Arrays.asList;
 import static net.bytebuddy.matcher.ElementMatchers.isConstructor;
-import static net.bytebuddy.matcher.ElementMatchers.isDeclaredBy;
 import static net.bytebuddy.matcher.ElementMatchers.isMethod;
 
 import com.amazonaws.services.sqs.model.ReceiveMessageRequest;
@@ -16,7 +13,7 @@ import net.bytebuddy.asm.Advice;
 
 @AutoService(Instrumenter.class)
 public class SqsReceiveRequestInstrumentation extends Instrumenter.Tracing
-    implements Instrumenter.ForKnownTypes {
+    implements Instrumenter.ForSingleType {
   public SqsReceiveRequestInstrumentation() {
     super("aws-sdk");
   }
@@ -27,26 +24,15 @@ public class SqsReceiveRequestInstrumentation extends Instrumenter.Tracing
   }
 
   @Override
-  public String[] knownMatchingTypes() {
-    return new String[] {
-      "com.amazonaws.services.sqs.model.ReceiveMessageRequest",
-      "com.amazonaws.services.sqs.buffered.QueueBufferConfig"
-    };
+  public String instrumentedType() {
+    return "com.amazonaws.services.sqs.model.ReceiveMessageRequest";
   }
 
   @Override
   public void adviceTransformations(AdviceTransformation transformation) {
     transformation.applyAdvice(
-        (isConstructor().or(isMethod().and(namedOneOf("setAttributeNames", "withAttributeNames"))))
-            .and(isDeclaredBy(named("com.amazonaws.services.sqs.model.ReceiveMessageRequest"))),
+        isConstructor().or(isMethod().and(namedOneOf("setAttributeNames", "withAttributeNames"))),
         getClass().getName() + "$ReceiveMessageRequestAdvice");
-    transformation.applyAdvice(
-        (isConstructor()
-                .or(
-                    isMethod()
-                        .and(namedOneOf("setReceiveAttributeNames", "withReceiveAttributeNames"))))
-            .and(isDeclaredBy(named("com.amazonaws.services.sqs.buffered.QueueBufferConfig"))),
-        getClass().getName() + "$QueueBufferConfigAdvice");
   }
 
   public static class ReceiveMessageRequestAdvice {
@@ -60,24 +46,6 @@ public class SqsReceiveRequestInstrumentation extends Instrumenter.Tracing
         }
       }
       attributeNames.add("AWSTraceHeader");
-    }
-  }
-
-  public static class QueueBufferConfigAdvice {
-    @Advice.OnMethodExit(suppress = Throwable.class)
-    public static void onExit(
-        @Advice.FieldValue(value = "receiveAttributeNames", readOnly = false)
-            List<String> receiveAttributeNames) {
-      // QueueBufferConfig maintains an immutable list which we may need to replace
-      for (String name : receiveAttributeNames) {
-        if ("AWSTraceHeader".equals(name) || "All".equals(name)) {
-          return;
-        }
-      }
-      int oldLength = receiveAttributeNames.size();
-      String[] nameArray = receiveAttributeNames.toArray(new String[oldLength + 1]);
-      nameArray[oldLength] = "AWSTraceHeader";
-      receiveAttributeNames = asList(nameArray);
     }
   }
 }
